Intro text change & other thoughts
I've targeted the language in the intro paragraph to be more for general users and citizen developers.  I had to look up "instantiate" in the dictionary, so I'm guessing many of our target audiences will too.  I also added more detail to the Basic Idea section, just because that's how I roll.

@sympmarc --What do you think about breaking out a separate page for **Universal Site Columns** and **Universal Content Types,** and then referring to that page rather than getting into the details on the Site Col & Content Type pages individually?  I think it would make it easier to describe them.
 
i.e.,  Universal Site Columns and Universal Content Types are created in a special hidden site called the "Content Type Hub" (*tenant*/sites/ContentTypeHub).  Creating a site column or Content type here can propagate it to all sites in your SharePoint environment.  
Sites must subscribe to the Content Type Hub in order for Universal Site Columns and Universal Content Types to be available to them.
Site Columns must be included in a propagated Universal Content Type for it to be included in the propagation--independent Site Columns that are not included in a propagated Universal Content Type are not delivered to other sites.
Changes made to a Universal Site Column or Universal Content Type can be pushed out to all sites that subscribe to it.  This push will update the replica that is available on the subscribed site, it will not change any existing columns or content types that are based upon it.

I think a link to my Metadata Ecosystem Infographic may help here, or using it as a visual...
https://straightenthemaze.com/intranets/technical-intranet/metadata/sharepoint-metadata-infographic/

I'd be happy to tackle the Universal SC/CT page

Cheers!  B