"How do they manage them" section updated
SharePoint Online UI allows to update the group privacy bit this statement was incorrect "For instance, to change a Group's Privacy policy you have to use Outlook or Outlook online. That setting can't be changed in SharePoint." 

It is updated as below with other example

For instance, to add guest users to office 365 group(both private and public) use Outlook or Outlook online. This setting is not present in SharePoint.